requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.9.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pandas>=2.0.0
tqdm>=4.66.0
colorama>=0.4.0